Fix 22.04 compiler error.

+++ src/gui/dialogs/dlg_options.cpp
@@ -597,7 +597,7 @@ OptionsDlg::OptionsDlg(wxWindow* parent, wxWindowID id, const wxString& title, c
     sbSizer_freedv700->Add(m_ckboxEnableLegacyModes, 0, static_cast<int>(wxALL) | wxALIGN_LEFT, 5);
 
     m_ckboxAutoStartOnLaunch = new wxCheckBox(sb_freedv700, wxID_ANY, _("Start Automatically on Launch"), wxDefaultPosition, wxDefaultSize, wxCHK_2STATE);
-    sbSizer_freedv700->Add(m_ckboxAutoStartOnLaunch, 0, wxALL | wxALIGN_LEFT, 5);
+    sbSizer_freedv700->Add(m_ckboxAutoStartOnLaunch, 0, static_cast<int>(wxALL) | wxALIGN_LEFT, 5);
 
     sizerModem->Add(sbSizer_freedv700, 0, static_cast<int>(wxALL)|static_cast<int>(wxEXPAND), 5);
